Fix Windows and mouse event crashes
Bumps LibreOfficeKit to v0.0.2 to fix some crashes on Windows.

Fixes cases where the access to the modifiers and position of a
WebMouseEvent is lost across function boundaries. Such cases cause a crash
that will not occur with inside of a Windows debugger (windbg or
Visual Studio) but will occur when run normally.

Fixes the case when the fonts missing event fires and is unhandled.
Now reports an unknown event instead of crashing when it is emitted.