Fix GPU rendering for scrolling and overlapping render errors
Since scroll events are now in the same task queue as invalidations, it
is guaranteed that the previous scroll event will occur before the next
paint. Paints however are not guaranteed, so on an occasion where paints
do not occur but a scroll invalidation is processed, the difference was 
not accounted for since the prior scroll position was stored at the end
of a paint. To avoid this, we use the previously stored scroll position
within the scroll update itself.