Disable partition alloc on all platforms to prevent crashes
There are several cases we've encountered where allocs and frees don't
match the correct allocator when going from Electron to LibreOfficeKit,
causing crashes. Usually an allocaiton is preformed in LOK with the
system default allocator but then Electron tries to free it with the
partition allocator. This is already disabled on macOS by default so
disabling it for Windows and Linux doesn't really change much.

They only just re-enabled partition alloc on Mac last week anyway:
https://github.com/electron/electron/pull/33981